Restless Fix
First IF statement runs restless if the player has maximum stamina, spending it and returning to not causing more to be gained. This means restless will now run if a player with max stamina turns it on, and the only other situation this would occur is if the player has not been online a long time, and the second IF statement has not run upon reaching max stamina as they theoretically could have earned more stamina. The second IF statement handles most cases, as it runs when the player gains 1 stamina, reaching max stamina and the next stamina they would earn is not due yet (Preventing them from spending the stamina when more could potentially still be earned from an offline backlog, which would lead to multiple uses if enough offline time had built up).